Rename VideoCard atom to VideoViewer as organism
